refactor: move wildcard dynamic parsing
